Partial support for array declaration and usage
Full support will need a code refactor since it is not possible the current state of the compiler. More  (like symbol table)  to be passed to the code generation in order to generate valid code. Also, I'm not sure how to handle array  using the [] operator. I need to do more research.